sched_exec should use zocl_xclbin to interact with xclbin.h
