+++ src/runtime_src/core/edge/drm/zocl/zocl_xclbin.c
@@ -942,3 +942,26 @@ zocl_xclbin_accel_adapter(int kds_mask)
 {
 	return kds_mask == ACCEL_ADAPTER;
 }
+
+/*
+ * returns false if any of the cu doesnt support interrupt
+ */
+bool
+zocl_xclbin_cus_support_intr(struct drm_zocl_dev *zdev)
+{
+	struct ip_data *ip;
+	int i;
+
+	if (!zdev->ip)
+		return false;
+
+	for (i = 0; i < zdev->ip->m_count; ++i) {
+		ip = &zdev->ip->m_ip_data[i];
+		if (!(ip->properties & 0x1)) {
+			return false;
+		}
+	}
+
+	return true;
+}
+
